Use the Inference Result in the Simulation

+++ src/str/rack_string.c
@@ -7,8 +7,8 @@
 
 void add_blanks(StringBuilder *string_builder, const Rack *rack,
                 const LetterDistribution *ld) {
-  int8_t number_of_blanks = rack_get_letter(rack, BLANK_MACHINE_LETTER);
-  for (int8_t j = 0; j < number_of_blanks; j++) {
+  uint16_t number_of_blanks = rack_get_letter(rack, BLANK_MACHINE_LETTER);
+  for (uint16_t j = 0; j < number_of_blanks; j++) {
     string_builder_add_user_visible_letter(string_builder, ld,
                                            BLANK_MACHINE_LETTER);
   }
@@ -21,8 +21,8 @@ void string_builder_add_rack(StringBuilder *string_builder, const Rack *rack,
   }
   for (int i = 0; i < rack_get_dist_size(rack); i++) {
     if (i != BLANK_MACHINE_LETTER) {
-      int8_t number_of_letter = rack_get_letter(rack, i);
-      for (int8_t j = 0; j < number_of_letter; j++) {
+      uint16_t number_of_letter = rack_get_letter(rack, i);
+      for (uint16_t j = 0; j < number_of_letter; j++) {
         string_builder_add_user_visible_letter(string_builder, ld, i);
       }
     }
